AS5311 Module Update
Changed again. Keep going back and forth between a ATtiny and ATmega IC. The ATtiny20 seemed like a good choice for the previous design, but it doesn't program via ISP, so I need to redo the board anyway. Price-wise an ATmega328 is almost the same, but much more headroom for future changes. IDK.